Use Proxy class to allow substitution of adapters
In order to be able to dynamically replace adapters being referenced by
calling modules, add a Proxy class. This class is a simple proxy for
Cache::Any adapters. It holds a reference to the adapter object, which
can be replaced at a later time. The calling module always retains
a reference to the same Proxy object.